// src/domain/enums.ts
export enum GameVariant {
  NL = "NL",
  KKD = "KKD",
  BE = "BE",
  INT = "INT",
  EUR = "EUR",
  INTERLANDS = "INTERLANDS",
}

export enum MarketType {
  ONE_X_TWO = "ONE_X_TWO",
  OVER_UNDER_25 = "OVER_UNDER_25",
  BOTH_TEAMS_TO_SCORE = "BOTH_TEAMS_TO_SCORE",
  CORRECT_SCORE = "CORRECT_SCORE",
}

export enum StrategyType {
  SAFE = "SAFE",
  BALANCED = "BALANCED",
  AGGRESSIVE = "AGGRESSIVE",
}